+++ sqry-daemon/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [28.0.1](https://github.com/verivus-oss/sqry/compare/v28.0.0...v28.0.1) - 2026-07-07
+
+### Other
+
+- update Cargo.lock dependencies
 ## [28.0.0](https://github.com/verivus-oss/sqry/compare/v27.0.8...v28.0.0) - 2026-07-07
 
 ### Added